\name{IDPmisc-package}
\alias{IDPmisc}
\docType{package}

\title{Utilities of Institute of Data Analyses and Process Design (www.idp.zhwin.ch)
}

\description{
  The IDPmisc package contains different high-level graphics functions
  for displaying large datasets, displaying circular data in a very
  flexible way, finding local maxima, brewing color ramps, drawing nice
  arrows, zooming 2D-plots, creating figures with differently colored
  margin and plot region.  In addition, the package contains auxiliary
  functions for data manipulation like omitting observations with
  irregular values or selecting data by logical vectors, which include
  NAs. Other functions are especially useful in spectroscopy and
  analyses of environmental data: robust baseline fitting, finding peaks
  in spectra.  
}

\details{
\tabular{ll}{
  Package: \tab IDPmisc\cr
  Type: \tab Package\cr
  Version: \tab 0.9.2\cr
  Date: \tab 2007-05-28\cr
  License: \tab  GPL (GNU Public Licence), Version 2 or later\cr
}

}

\author{
Rene Locher, Andreas Ruckstuhl, Thomas Unternährer

Maintainer: Rene Locher <rene.locher@zhwin.ch>
}

%% references{
%% ~ Literature or other references for background information ~~
%%

\keyword{ package }
\keyword{ dplot }
\keyword{ hplot }
\keyword{ manip }
\keyword{ utilities }
\keyword{ ts }
\keyword{ multivariate }

%% \seealso{
%% ~~ Optional links to other man pages, e.g. ~~
%% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%% }
%% 
%% \examples{
%% ~~ simple examples of the most important functions ~~
%% }
