\name{ixyplot}
\alias{ixyplot}


\concept{large dataset}
\concept{huge dataset}

\title{Image Scatter Plot for Large Datasets}

\description{
  Defunct. Use \code{iplot()} instead!
}

\details{
  Defunct. Use \code{iplot()} instead!
}

\author{Andreas Ruckstuhl, Rene Locher}


\seealso{\code{\link{ilagplot}}, \code{\link{ipairs}}, \code{\link{Image}}}

\keyword{hplot}
