\name{internals}
\alias{h.image}
\alias{h.LegendAndTitle}

\title{Internal Functions}
\description{
  These functions are for internal use only and not documented further.
}
\usage{
h.image(x, y, pixs = 1, zmax = NULL, colramp = IDPcolorRamp)
h.LegendAndTitle(main, cex.main, border, colramp, zmax)
}

\author{Rene Locher}

\keyword{internal}

