\name{rfbaselineScale}
\alias{rfbaselineScale}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Estimation of the Scale Parameter }
\description{
  Estimation of the scale parameter based on data smaller than its first
  mode. Mainly used in rfbaseline.
}
\usage{
rfbaselineScale(r)
}

\arguments{
  \item{r}{ residuals }
}

\value{
Estimated scale.
}

\author{ Andreas Ruckstuhl}

\seealso{ See also \code{\link{mad}} }

\keyword{ univar }
