\name{IDPmisc-package}
\alias{IDPmisc-package}
\alias{IDPmisc}
\docType{package}

\title{Utilities of Institute of Data Analyses and Process Design (www.zhaw.ch/idp)
}

\description{
  Different high-level graphics functions for displaying large datasets, displaying circular data in a very flexible way, finding local maxima, brewing color ramps, drawing nice arrows, zooming 2D-plots, creating figures with differently colored margin and plot region.  In addition, the package contains auxiliary functions for data manipulation like omitting observations with irregular values or selecting data by logical vectors, which include NAs. Other functions are especially useful in spectroscopy and analyses of environmental data: robust baseline fitting, finding peaks in spectra, converting humidity measures.
}

\author{
Rene Locher, Andreas Ruckstuhl et al.
Maintainer: Christoph Hofer <christoph.hofer@zhaw.ch>
}

%% references{
%% ~ Literature or other references for background information ~~
%%

\keyword{ package }
\keyword{ arith }
\keyword{ dplot }
\keyword{ hplot }
\keyword{ manip }
\keyword{ utilities }
\keyword{ ts }
\keyword{ multivariate }

%% \seealso{
%% ~~ Optional links to other man pages, e.g. ~~
%% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%% }
%%
%% \examples{
%% ~~ simple examples of the most important functions ~~
%% }
