\name{internals}
\alias{LegendAndTitle}

\title{Internal Functions}
\description{
  These functions are for internal use only and not documented further.
}
\usage{
LegendAndTitle(main, cex.main, border, colramp, zmax)
plotdat(rho, cyclVar, circle, transf, vp, general, grid, title)
rose.grob(pdat, general, grid, title)
}

\details{
  \tabular{ll}{
    Function \tab used in\cr
    \code{LegendAndTitle}
    \tab \code{\link{ipairs}}, \code{\link{ilagplot}}\cr
    \code{plotdat} \tab \code{\link{plot.rose}}\cr
    \code{rose.grob} \tab \code{\link{plot.rose}}\cr
    \code{} \tab \code{\link{plot.rose}}\cr
  }
  
}

\author{Rene Locher}

\keyword{internal}

