\name{internals}
\alias{h.LegendAndTitle}
\alias{h.inf.omit}

\title{Internal Functions}
\description{
  These functions are for internal use only and not documented further.
}
\usage{
h.LegendAndTitle(main, cex.main, border, colramp, zmax)
h.inf.omit(x)
}

\author{Rene Locher}

\keyword{internal}

